+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Groups rapid fixes and improvements [#447](https://github.com/rokwire/groups-building-block/issues/447)
+- Group Create Post adds members by default [#442](https://github.com/rokwire/groups-building-block/issues/442)
+- Implement create and update post admin APIs [#448](https://github.com/rokwire/groups-building-block/issues/448)
+- API for creating Group for the Admin app [#445](https://github.com/rokwire/groups-building-block/issues/445)
+- API for updating group for the Admin app [#446](https://github.com/rokwire/groups-building-block/issues/446)
+- Ability to filter authman/manged groups [#441](https://github.com/rokwire/groups-building-block/issues/441)
+
 ## [1.37.1] - 2024-04-24
 ### Fixed
 - Additional fix for scheduled posts[#437](https://github.com/rokwire/groups-building-block/issues/437)